refactor(Transaction): void 반환 타입 미사용에 따른 제거
