--- src/main/java/janggi/infra/VoidTransactionCallback.java
@@ -1,9 +1,0 @@
-package janggi.infra;
-
-import java.sql.Connection;
-import java.sql.SQLException;
-
-@FunctionalInterface
-public interface VoidTransactionCallback {
-    void doInTransaction(Connection connection) throws SQLException;
-}
