+++ src/main/java/janggi/infra/transaction/TransactionTemplate.java
@@ -1,5 +1,6 @@
-package janggi.infra;
+package janggi.infra.transaction;
 
+import janggi.infra.datasource.DataConnectionManager;
 import java.sql.Connection;
 import java.sql.SQLException;
 
@@ -19,36 +20,12 @@ public class TransactionTemplate {
         }
     }
 
-    public void executeInTransaction(VoidTransactionCallback action) {
-        try (Connection connection = manager.getConnection()) {
-            processTransaction(connection, action);
-        } catch (SQLException e) {
-            throw new RuntimeException("[ERROR] DB 커넥션 에러", e);
-        }
-    }
-
     private <T> T processTransaction(Connection connection, TransactionCallback<T> action) throws SQLException {
         try {
             connection.setAutoCommit(false);
             T result = action.doInTransaction(connection);
             connection.commit();
             return result;
-        } catch (RuntimeException e) {
-            connection.rollback();
-            throw e;
-        } catch (SQLException e) {
-            connection.rollback();
-            throw new RuntimeException("[ERROR] 게임 저장 중 트랜잭션 롤백됨", e);
-        } finally {
-            connection.setAutoCommit(true);
-        }
-    }
-
-    private void processTransaction(Connection connection, VoidTransactionCallback action) throws SQLException {
-        try {
-            connection.setAutoCommit(false);
-            action.doInTransaction(connection);
-            connection.commit();
         } catch (RuntimeException e) {
             connection.rollback();
             throw e;
